requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.2.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
